app/requirements.txt
asgiref==3.5.2
django-ckeditor==6.5.0
django-js-asset==2.0.0
Django==4.1
Pillow==9.2.0
pip==22.2.2
setuptools==65.3.0
sqlparse==0.4.2
wheel==0.37.1
gunicorn==20.1.0
psycopg2-binary==2.9.5
